refactor hooks
