+++ src/App.tsx
@@ -26,7 +26,7 @@ import { KeyringContextProvider } from './contexts/KeyringContextProvider';
 import { MessageContextProvider } from './contexts/MessageContext';
 import { SourceTargetContextProvider } from './contexts/SourceTargetContextProvider';
 import { TransactionContextProvider } from './contexts/TransactionContext';
-import { useApiConnection } from './hooks/useApiConnection';
+import { useApiConnection } from './hooks';
 import Main from './screens/Main';
 
 const providers = getChainProviders();
