+++ src/components/CustomCall.tsx
@@ -23,8 +23,7 @@ import styled from 'styled-components';
 
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useTransactionContext } from '../contexts/TransactionContext';
-import useLoadingApi from '../hooks/useLoadingApi';
-import useSendMessage from '../hooks/useSendMessage';
+import { useLoadingApi, useSendMessage } from '../hooks';
 import { TransactionTypes } from '../types/transactionTypes';
 interface Props {
   className?: string;
