+++ src/components/Transfer.tsx
@@ -21,9 +21,8 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 
 import { useTransactionContext } from '../contexts/TransactionContext';
-import useConnectedReceiver from '../hooks/useConnectedReceiver';
-import useLoadingApi from '../hooks/useLoadingApi';
-import useSendMessage from '../hooks/useSendMessage';
+import { useConnectedReceiver } from '../hooks';
+import { useLoadingApi, useSendMessage } from '../hooks';
 import { TransactionTypes } from '../types/transactionTypes';
 
 interface Props {
