+++ src/hooks/useSendMessage.ts
@@ -25,8 +25,7 @@ import { useAccountContext } from '../contexts/AccountContextProvider';
 import { useUpdateMessageContext } from '../contexts/MessageContext';
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useTransactionContext, useUpdateTransactionContext } from '../contexts/TransactionContext';
-import useLaneId from '../hooks/useLaneId';
-import useTransactionPreparation from '../hooks/useTransactionPreparation';
+import { useLaneId, useTransactionPreparation } from '../hooks';
 import { TransactionStatusEnum, TransactionTypes } from '../types/transactionTypes';
 import getSubstrateDynamicNames from '../util/getSubstrateDynamicNames';
 import logger from '../util/logger';
