+++ src/hooks/useTransactionPreparation.ts
@@ -21,9 +21,7 @@ import { TransactionActionCreators } from '../actions/transactionActions';
 import { useAccountContext } from '../contexts/AccountContextProvider';
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useUpdateTransactionContext } from '../contexts/TransactionContext';
-import useLaneId from '../hooks/useLaneId';
-import useLoadingApi from '../hooks/useLoadingApi';
-import useTransactionType from '../hooks/useTransactionType';
+import { useLaneId, useLoadingApi, useTransactionType } from '../hooks';
 import getSubstrateDynamicNames from '../util/getSubstrateDynamicNames';
 
 interface Props {
